Create extensions files
