+++ app/src/main/java/com/slg/pokeonary/mobile/common/ViewGroupExtensions.kt
@@ -1,0 +1,11 @@
+package com.slg.pokeonary.mobile.common
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+fun ViewGroup.inflate(layoutId: Int, attachToRoot: Boolean = false): View =
+    LayoutInflater.from(context).inflate(
+        layoutId, this,
+        attachToRoot
+    )
